requirements.txt
astroid==4.0.2
black==25.11.0
cachetools==6.2.2
chardet==5.2.0
click==8.3.1
colorama==0.4.6
coverage==7.12.0
dill==0.4.0
distlib==0.4.0
filelock==3.20.0
flake8==7.3.0
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.4
pyproject-api==1.10.0
pytest==9.0.1
pytest-cov==7.0.0
pytokens==0.3.0
PyYAML==6.0.3
tomlkit==0.13.3
tox==4.32.0
virtualenv==20.35.4
